requirements.txt
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
idna==3.11
oracledb==3.4.1
pycparser==2.23
python-dotenv==1.2.1
requests==2.32.5
typing_extensions==4.15.0
urllib3==2.5.0
